fix errors in MSVC compile ( can't use or and not as keywords, when making a prototype conts, make the implementation const as well) fix constructor warnings (arrays don't have default inits for some reason) fix cast warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22200 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ plugins/TimeLimit/TimeLimit.cpp
@@ -182,7 +182,7 @@ bool TimeLimit::SlashCommand ( int playerID, bz_ApiString cmd, bz_ApiString, bz_
   }
 
   if (nonumber == 0 ) {
-    float limit = atof(cmdParams->get(0).c_str());
+    float limit = (float)atof(cmdParams->get(0).c_str());
 	// Don't allow timelimit being equal or lower then 0
 	if (limit > 0 ) {
